+++ packages/js/api/src/models/products/variable-product.ts
@@ -115,10 +115,11 @@ export class VariableProduct extends AbstractProduct implements
 	public readonly onePerOrder: boolean = false;
 	public readonly trackInventory: boolean = false;
 	public readonly remainingStock: number = -1;
-	public readonly stockStatus: StockStatus = ''
+	public readonly stockStatus: StockStatus = '';
 	public readonly backorderStatus: BackorderStatus = BackorderStatus.Allowed;
 	public readonly canBackorder: boolean = false;
 	public readonly isOnBackorder: boolean = false;
+	public readonly lowStockThreshold: number = -1;
 
 	/**
 	 * @see ./abstracts/sales-tax.ts