+++ packages/js/e2e-environment/bin/docker-compose.js
@@ -10,6 +10,8 @@ const {
 	getAppRoot,
 	getAppName,
 	getTestConfig,
+	resolveLocalE2ePath,
+	resolvePackagePath,
 } = require( '../utils' );
 
 const dockerArgs = [];
@@ -56,16 +58,13 @@ if ( appPath ) {
 
 		const appInitFile = customInitFile
 			? customInitFile
-			: path.resolve(
-					appPath,
-					'plugins/woocommerce/tests/e2e/docker/initialize.sh'
-			  );
+			: resolveLocalE2ePath( 'docker/initialize.sh' );
 		// If found, copy it into the wp-cli Docker context so
 		// it gets picked up by the entrypoint script.
 		if ( fs.existsSync( appInitFile ) ) {
 			fs.copyFileSync(
 				appInitFile,
-				path.resolve( __dirname, '../docker/wp-cli/initialize.sh' )
+				resolvePackagePath( 'docker/wp-cli/initialize.sh' )
 			);
 			console.log( 'Initializing ' + appInitFile );
 		}
@@ -84,11 +83,6 @@ if ( ! process.env.WC_E2E_FOLDER_MAPPING ) {
 		'/var/www/html/wp-content/plugins/' + getAppBase();
 }
 
-// Set some environment variables
-if ( ! process.env.WC_CORE_PATH ) {
-	envVars.WC_CORE_PATH = 'plugins/woocommerce';
-}
-
 if ( ! process.env.WORDPRESS_PORT ) {
 	process.env.WORDPRESS_PORT = testConfig.port;
 }
@@ -97,7 +91,7 @@ if ( ! process.env.WORDPRESS_URL ) {
 }
 
 // Ensure that the first Docker compose file loaded is from our local env.
-dockerArgs.unshift( '-f', path.resolve( __dirname, '../docker-compose.yaml' ) );
+dockerArgs.unshift( '-f', resolvePackagePath( 'docker-compose.yaml' ) );
 
 const dockerProcess = spawnSync( 'docker-compose', dockerArgs, {
 	stdio: 'inherit',