+++ packages/js/e2e-utils/src/flows/with-rest-api.js
@@ -4,12 +4,14 @@ import {Coupon, Setting, SimpleProduct, Order} from '@woocommerce/api';
 
 const client = factories.api.withDefaultPermalinks;
 const onboardingProfileEndpoint = '/wc-admin/onboarding/profile';
-const shippingZoneEndpoint = '/wc/v3/shipping/zones';
-const shippingClassesEndpoint = '/wc/v3/products/shipping_classes';
-const userEndpoint = '/wp/v2/users';
-const systemStatusEndpoint = '/wc/v3/system_status';
 const productsEndpoint = '/wc/v3/products';
 const productCategoriesEndpoint = '/wc/v3/products/categories';
+const shippingClassesEndpoint = '/wc/v3/products/shipping_classes';
+const shippingZoneEndpoint = '/wc/v3/shipping/zones';
+const systemStatusEndpoint = '/wc/v3/system_status';
+const taxClassesEndpoint = '/wc/v3/taxes/classes';
+const taxRatesEndpoint = '/wc/v3/taxes';
+const userEndpoint = '/wp/v2/users';
 
 /**
  * Utility function to delete all merchant created data store objects.
@@ -43,6 +45,16 @@ const deleteAllRepositoryObjects = async ( repository, defaultObjectId = null, s
 };
 
 /**
+ * Utility to flatten a tax rate.
+ *
+ * @param {object} taxRate Tax rate to be flattened.
+ * @return {string}
+ */
+const flattenTaxRate = ( taxRate ) => {
+	return taxRate.rate + '/' + taxRate.class + '/' + taxRate.name;
+};
+
+/**
  * Utility functions that use the REST API to process the requested function.
  */
 export const withRestApi = {
@@ -66,7 +78,7 @@ export const withRestApi = {
 		};
 
 		const response = await client.put( onboardingProfileEndpoint, onboardingReset );
-		expect( response.status ).toEqual( 200 );
+		expect( response.statusCode ).toEqual( 200 );
 	},
 	/**
 	 * Use api package to delete coupons.
@@ -76,6 +88,16 @@ export const withRestApi = {
 	deleteAllCoupons: async () => {
 		const repository = Coupon.restRepository( client );
 		await deleteAllRepositoryObjects( repository );
+	},
+	/**
+	 * Use api package to delete a coupon.
+	 *
+	 * @param {number} couponId Coupon ID.
+	 * @return {Promise} Promise resolving once coupon has been deleted.
+	 */
+	deleteCoupon: async ( couponId ) => {
+		const repository = Coupon.restRepository( client );
+		await repository.delete( couponId );
 	},
 	/**
 	 * Use api package to delete products.
@@ -89,49 +111,58 @@ export const withRestApi = {
 	/**
 	 * Use the API to delete all product attributes.
 	 *
+	 * @param {boolean} testResponse Test the response status code.
 	 * @return {Promise} Promise resolving once attributes have been deleted.
 	 */
-	deleteAllProductAttributes: async () => {
+	deleteAllProductAttributes: async ( testResponse = true ) => {
 		const productAttributesPath = productsEndpoint + '/attributes';
 		const productAttributes = await client.get( productAttributesPath );
 		if ( productAttributes.data && productAttributes.data.length ) {
 			for ( let a = 0; a < productAttributes.data.length; a++ ) {
 				const response = await client.delete( productAttributesPath + `/${productAttributes.data[a].id}?force=true` );
-				expect( response.status ).toBe( 200 );
+				if ( testResponse ) {
+					expect( response.status ).toBe( 200 );
+				}
 			}
 		}
 	},
 	/**
 	 * Use the API to delete all product categories.
 	 *
+	 * @param {boolean} testResponse Test the response status code.
 	 * @return {Promise} Promise resolving once categories have been deleted.
 	 */
-	deleteAllProductCategories: async () => {
+	deleteAllProductCategories: async ( testResponse = true ) => {
 		const productCategoriesPath = productsEndpoint + '/categories';
 		const productCategories = await client.get( productCategoriesPath );
 		if ( productCategories.data && productCategories.data.length ) {
 			for ( let c = 0; c < productCategories.data.length; c++ ) {
 				// The default `uncategorized` category can't be deleted
-				if ( productCategories.data[c].id == 0 ) {
+				if ( productCategories.data[c].slug == 'uncategorized' ) {
 					continue;
 				}
 				const response = await client.delete( productCategoriesPath + `/${productCategories.data[c].id}?force=true` );
-				expect( response.status ).toBe( 200 );
+				if ( testResponse ) {
+					expect( response.status ).toBe( 200 );
+				}
 			}
 		}
 	},
 	/**
 	 * Use the API to delete all product tags.
 	 *
+	 * @param {boolean} testResponse Test the response status code.
 	 * @return {Promise} Promise resolving once tags have been deleted.
 	 */
-	deleteAllProductTags: async () => {
+	deleteAllProductTags: async ( testResponse = true ) => {
 		const productTagsPath = productsEndpoint + '/tags';
 		const productTags = await client.get( productTagsPath );
 		if ( productTags.data && productTags.data.length ) {
 			for ( let t = 0; t < productTags.data.length; t++ ) {
 				const response = await client.delete( productTagsPath + `/${productTags.data[t].id}?force=true` );
-				expect( response.status ).toBe( 200 );
+				if ( testResponse ) {
+					expect( response.status ).toBe( 200 );
+				}
 			}
 		}
 	},
@@ -155,6 +186,7 @@ export const withRestApi = {
 	 * @param zoneMethod Shipping method type. Defaults to flat_rate (use also: free_shipping or local_pickup).
 	 * @param cost Shipping method cost. Default is no cost.
 	 * @param additionalZoneMethods Array of additional zone methods to add to the shipping zone.
+	 * @param {boolean} testResponse Test the response status code.
 	 */
 	addShippingZoneAndMethod: async (
 		zoneName,
@@ -162,33 +194,38 @@ export const withRestApi = {
 		zipCode = '',
 		zoneMethod = 'flat_rate',
 		cost = '',
-		additionalZoneMethods = [] ) => {
+		additionalZoneMethods = [],
+		testResponse = true ) => {
 
-	   const path = 'wc/v3/shipping/zones';
+		const path = 'wc/v3/shipping/zones';
 
-	   const response = await client.post( path, { name: zoneName } );
-	   expect(response.status).toEqual(201);
-	   let zoneId = response.data.id;
+		const response = await client.post( path, { name: zoneName } );
+		if ( testResponse ) {
+			expect( response.status ).toEqual( 201 );
+		}
+		let zoneId = response.data.id;
 
-	   // Select shipping zone location
-	   let [ zoneType, zoneCode ] = zoneLocation.split(/:(.+)/);
-	   let zoneLocationPayload = [
+		// Select shipping zone location
+		let [ zoneType, zoneCode ] = zoneLocation.split(/:(.+)/);
+		let zoneLocationPayload = [
 		   {
 			   code: zoneCode,
 			   type: zoneType,
 		   }
-	   ];
+		];
 
-	   // Fill shipping zone postcode if provided
-	   if ( zipCode ) {
+		// Fill shipping zone postcode if provided
+		if ( zipCode ) {
 		   zoneLocationPayload.push( {
 			   code: zipCode,
 			   type: "postcode",
 		   } );
-	   }
+		}
 
-	   const locationResponse = await client.put( path + `/${zoneId}/locations`, zoneLocationPayload );
-	   expect(locationResponse.status).toEqual(200);
+		const locationResponse = await client.put( path + `/${zoneId}/locations`, zoneLocationPayload );
+		if ( testResponse ) {
+			expect( locationResponse.status ).toEqual( 200 );
+		}
 
 	   // Add shipping zone method
 	   let methodPayload = {
@@ -196,7 +233,9 @@ export const withRestApi = {
 	   }
 
 	   const methodsResponse = await client.post( path + `/${zoneId}/methods`, methodPayload );
-	   expect(methodsResponse.status).toEqual(200);
+		if ( testResponse ) {
+			expect( methodsResponse.status ).toEqual( 200 );
+		}
 	   let methodId = methodsResponse.data.id;
 
 	   // Add in cost, if provided
@@ -208,23 +247,28 @@ export const withRestApi = {
 		   }
 
 		   const costResponse = await client.put( path + `/${zoneId}/methods/${methodId}`, costPayload );
-		   expect(costResponse.status).toEqual(200);
+		   if ( testResponse ) {
+			   expect( costResponse.status ).toEqual( 200 );
+		   }
 	   }
 
 	   // Add any additional zones, if provided
 	   if (additionalZoneMethods.length > 0) {
 		   for ( let z = 0; z < additionalZoneMethods.length; z++ ) {
 			   let response = await client.post( path + `/${zoneId}/methods`, { method_id: additionalZoneMethods[z] } );
-			   expect(response.status).toEqual(200);
+			   if ( testResponse ) {
+				   expect( response.status ).toBe( 200 );
+			   }
 		   }
 	   }
     },
 	/**
 	 * Use api package to delete shipping zones.
 	 *
+	 * @param {boolean} testResponse Test the response status code.
 	 * @return {Promise} Promise resolving once shipping zones have been deleted.
 	 */
-	deleteAllShippingZones: async () => {
+	deleteAllShippingZones: async ( testResponse = true ) => {
 		const shippingZones = await client.get( shippingZoneEndpoint );
 		if ( shippingZones.data && shippingZones.data.length ) {
 			for ( let z = 0; z < shippingZones.data.length; z++ ) {
@@ -233,21 +277,26 @@ export const withRestApi = {
 					continue;
 				}
 				const response = await client.delete( shippingZoneEndpoint + `/${shippingZones.data[z].id}?force=true` );
-				expect( response.status ).toBe( 200 );
+				if ( testResponse ) {
+					expect( response.status ).toBe( 200 );
+				}
 			}
 		}
 	},
 	/**
 	 * Use api package to delete shipping classes.
 	 *
+	 * @param {boolean} testResponse Test the response status code.
 	 * @return {Promise} Promise resolving once shipping classes have been deleted.
 	 */
-	deleteAllShippingClasses: async () => {
+	deleteAllShippingClasses: async ( testResponse = true ) => {
 		const shippingClasses = await client.get( shippingClassesEndpoint );
 		if ( shippingClasses.data && shippingClasses.data.length ) {
 			for ( let c = 0; c < shippingClasses.data.length; c++ ) {
 				const response = await client.delete( shippingClassesEndpoint + `/${shippingClasses.data[c].id}?force=true` );
-				expect( response.status ).toBe( 200 );
+				if ( testResponse ) {
+					expect( response.status ).toBe( 200 );
+				}
 			}
 		}
 	},
@@ -278,9 +327,10 @@ export const withRestApi = {
 	/**
 	 * Reset a settings group to default values except selects.
 	 * @param settingsGroup
+	 * @param {boolean} testResponse Test the response status code.
 	 * @returns {Promise<void>}
 	 */
-	resetSettingsGroupToDefault: async ( settingsGroup ) => {
+	resetSettingsGroupToDefault: async ( settingsGroup, testResponse = true ) => {
 		const settingsClient = Setting.restRepository( client );
 		const settings = await settingsClient.list( settingsGroup );
 		if ( ! settings.length  ) {
@@ -300,7 +350,7 @@ export const withRestApi = {
 
 			const response = await settingsClient.update( settingsGroup, defaultSetting.id, defaultSetting );
 			// Multi-selects have a default '' but return an empty [].
-			if ( settings[s].type != 'multiselect' ) {
+			if ( testResponse && settings[s].type != 'multiselect' ) {
 				expect( response.value ).toBe( defaultSetting.value );
 			}
 		}
@@ -321,22 +371,61 @@ export const withRestApi = {
 	 *
 	 * @param {string} paymentGatewayId The ID of the payment gateway to update.
 	 * @param {object} payload An object with the key/value pair to update.
+	 * @param {boolean} testResponse Test the response status code.
 	 */
-	updatePaymentGateway: async ( paymentGatewayId, payload = {} ) => {
+	updatePaymentGateway: async ( paymentGatewayId, payload = {}, testResponse = true ) => {
 		const response = await client.put( `/wc/v3/payment_gateways/${paymentGatewayId}`, payload );
-		expect( response.status ).toBe( 200 );
+		if ( testResponse ) {
+			expect( response.status ).toEqual( 200 );
+		}
 	},
 	/**
 	 * Create a batch of orders using the "Batch Create Order" API endpoint.
 	 *
 	 * @param orders Array of orders to be created
+	 * @param {boolean} testResponse Test the response status code.
 	 */
-	batchCreateOrders: async (orders) => {
+	batchCreateOrders: async ( orders, testResponse = true ) => {
 		const path = '/wc/v3/orders/batch';
 		const payload = { create: orders };
 
 		const response = await client.post(path, payload);
-		expect( response.status ).toEqual(200);
+		if ( testResponse ) {
+			expect( response.status ).toBe( 200 );
+		}
+	},
+	/**
+	 * Add tax classes.
+	 *
+	 * @param {<Array<Object>>} taxClasses Array of tax class objects.
+	 * @returns {Promise<void>}
+	 */
+	addTaxClasses: async ( taxClasses ) => {
+		// Only add tax classes which don't already exist.
+		const existingTaxClasses = await client.get( taxClassesEndpoint );
+		const existingTaxNames = existingTaxClasses.data.map( taxClass => taxClass.name );
+		const newTaxClasses = taxClasses.filter( taxClass => ! existingTaxNames.includes( taxClass.name ) );
+
+		for ( const taxClass of newTaxClasses ) {
+			await client.post( taxClassesEndpoint, taxClass );
+		}
+	},
+	/**
+	 * Add tax rates.
+	 *
+	 * @param {<Array<Object>>} taxRates Array of tax rate objects.
+	 * @returns {Promise<void>}
+	 */
+	addTaxRates: async ( taxRates ) => {
+		// Only add rates which don't already exist
+		const existingTaxRates = await client.get( taxRatesEndpoint );
+		const existingRates = existingTaxRates.data.map( taxRate => flattenTaxRate( taxRate ) );
+
+		for ( const taxRate of taxRates ) {
+			if ( ! existingRates.includes( flattenTaxRate( taxRate ) ) ) {
+				await client.post( taxRatesEndpoint, taxRate );
+			}
+		}
 	},
 	/**
 	 * Get the current environment from the WooCommerce system status API.