+++ plugins/woocommerce/includes/abstracts/abstract-wc-order.php
@@ -1166,7 +1166,12 @@ abstract class WC_Abstract_Order extends WC_Abstract_Legacy_Order {
 			$used_by = $this->get_billing_email();
 		}
 
-		$coupon->increase_usage_count( $used_by );
+		$order_data_store = $this->get_data_store();
+		if ( $order_data_store->get_recorded_coupon_usage_counts( $this ) ) {
+			$coupon->increase_usage_count( $used_by );
+		}
+
+		wc_update_coupon_usage_counts( $this->get_id() );
 
 		return true;
 	}
@@ -1318,6 +1323,16 @@ abstract class WC_Abstract_Order extends WC_Abstract_Legacy_Order {
 				if ( ! $item_id ) {
 					$coupon_item = new WC_Order_Item_Coupon();
 					$coupon_item->set_code( $coupon_code );
+
+					// Add coupon data.
+					$coupon_id = wc_get_coupon_id_by_code( $coupon_code );
+					$coupon    = new WC_Coupon( $coupon_id );
+
+					// Avoid storing used_by - it's not needed and can get large.
+					$coupon_data = $coupon->get_data();
+					unset( $coupon_data['used_by'] );
+
+					$coupon_item->add_meta_data( 'coupon_data', $coupon_data );
 				} else {
 					$coupon_item = $this->get_item( $item_id, false );
 				}