+++ plugins/woocommerce/includes/admin/class-wc-admin-addons.php
@@ -87,17 +87,57 @@ class WC_Admin_Addons {
 				)
 			);
 
-			if ( ! is_wp_error( $raw_featured ) ) {
-				$featured = json_decode( wp_remote_retrieve_body( $raw_featured ) );
-				if ( $featured ) {
-					set_transient( 'wc_addons_featured', $featured, DAY_IN_SECONDS );
-				}
+			if ( is_wp_error( $raw_featured ) ) {
+				do_action( 'woocommerce_page_wc-addons_connection_error', $raw_featured->get_error_message() );
+
+				$message = self::is_ssl_error( $raw_featured->get_error_message() )
+					? __( 'We encountered an SSL error. Please ensure your site supports TLS version 1.2 or above.', 'woocommerce' )
+					: $raw_featured->get_error_message();
+
+				self::output_empty( $message );
+
+				return;
+			}
+
+			$response_code = (int) wp_remote_retrieve_response_code( $raw_featured );
+			if ( 200 !== $response_code ) {
+				do_action( 'woocommerce_page_wc-addons_connection_error', $response_code );
+
+				/* translators: %d: HTTP error code. */
+				$message = sprintf(
+					esc_html(
+						__(
+							'Our request to the featured API got error code %d.',
+							'woocommerce'
+						)
+					),
+					$response_code
+				);
+
+				self::output_empty( $message );
+
+				return;
+			}
+
+			$featured      = json_decode( wp_remote_retrieve_body( $raw_featured ) );
+			if ( empty( $featured ) || ! is_array( $featured ) ) {
+				do_action( 'woocommerce_page_wc-addons_connection_error', 'Empty or malformed response' );
+				$message = __( 'Our request to the featured API got a malformed response.', 'woocommerce' );
+				self::output_empty( $message );
+
+				return;
+			}
+
+			if ( $featured ) {
+				set_transient( 'wc_addons_featured', $featured, DAY_IN_SECONDS );
 			}
 		}
 
-		if ( ! empty( $featured ) ) {
-			self::output_featured( $featured );
-		}
+		self::output_featured( $featured );
+	}
+
+	public static function is_ssl_error( $error_message ) {
+		return false !== stripos( $error_message, 'cURL error 35' );
 	}
 
 	/**
@@ -127,7 +167,7 @@ class WC_Admin_Addons {
 	 * @param  string $term     Search terms.
 	 * @param  string $country  Store country.
 	 *
-	 * @return object of extensions and promotions.
+	 * @return object|WP_Error  Object with products and promotions properties, or WP_Error
 	 */
 	public static function get_extension_data( $category, $term, $country ) {
 		$parameters = self::build_parameter_string( $category, $term, $country );
@@ -144,9 +184,24 @@ class WC_Admin_Addons {
 			array( 'headers' => $headers )
 		);
 
-		if ( ! is_wp_error( $raw_extensions ) ) {
-			$addons = json_decode( wp_remote_retrieve_body( $raw_extensions ) );
+		if ( is_wp_error( $raw_extensions ) ) {
+			do_action( 'woocommerce_page_wc-addons_connection_error', $raw_extensions->get_error_message() );
+			return $raw_extensions;
 		}
+
+		$response_code = (int) wp_remote_retrieve_response_code( $raw_extensions );
+		if ( 200 !== $response_code ) {
+			do_action( 'woocommerce_page_wc-addons_connection_error', $response_code );
+			return new WP_Error( 'error', __( "Our request to the search API got response code $response_code.", 'woocommerce' ) );
+		}
+
+		$addons = json_decode( wp_remote_retrieve_body( $raw_extensions ) );
+
+		if ( ! is_object( $addons ) || ! isset( $addons->products ) ) {
+			do_action( 'woocommerce_page_wc-addons_connection_error', 'Empty or malformed response' );
+			return new WP_Error( 'error', __( "Our request to the search API got a malformed response.", 'woocommerce' ) );
+		}
+
 		return $addons;
 	}
 
@@ -758,7 +813,7 @@ class WC_Admin_Addons {
 		$product_list_classes = 'products addons-products-' . $product_list_classes;
 		?>
 			<section class="addon-product-group">
-				<h1 class="addon-product-group-title"><?php echo esc_html( $block->title ); ?></h1>
+				<h2 class="addon-product-group-title"><?php echo esc_html( $block->title ); ?></h2>
 				<div class="addon-product-group-description-container">
 					<?php if ( ! empty( $block->description ) ) : ?>
 					<div class="addon-product-group-description">
@@ -908,6 +963,31 @@ class WC_Admin_Addons {
 		<?php
 	}
 
+	public static function output_empty( $message = '' ) {
+		?>
+		<div class="wc-addons__empty">
+			<h2><?php echo wp_kses_post( __( 'Oh no! We\'re having trouble connecting to the extensions catalog right now.', 'woocommerce' ) ); ?></h2>
+			<?php if ( ! empty( $message ) ) : ?>
+				<p><?php echo esc_html( $message ); ?></p>
+			<?php endif; ?>
+			<p>
+				<?php
+				/* translators: a url */
+				printf(
+					wp_kses_post(
+						__(
+							'To start growing your business, head over to <a href="%s">WooCommerce.com</a>, where you\'ll find the most popular WooCommerce extensions.',
+							'woocommerce'
+						)
+					),
+					'https://woocommerce.com/products/?utm_source=extensionsscreen&utm_medium=product&utm_campaign=connectionerror'
+				);
+				?>
+			</p>
+		</div>
+		<?php
+	}
+
 
 	/**
 	 * Handles output of the addons page in admin.
@@ -946,7 +1026,7 @@ class WC_Admin_Addons {
 			$term           = $search ? $search : null;
 			$country        = WC()->countries->get_base_country();
 			$extension_data = self::get_extension_data( $category, $term, $country );
-			$addons         = $extension_data->products;
+			$addons         = is_wp_error( $extension_data ) ? $extension_data : $extension_data->products;
 			$promotions     = ! empty( $extension_data->promotions ) ? $extension_data->promotions : array();
 		}
 
@@ -1161,28 +1241,34 @@ class WC_Admin_Addons {
 			// For product-related banners icon is a product's image.
 			$mapped->icon = $data->image ?? null;
 		}
+
 		// URL.
 		$mapped->url = $data->link ?? null;
 		if ( empty( $mapped->url ) ) {
 			$mapped->url = $data->url ?? null;
 		}
+
 		// Title.
 		$mapped->title = $data->title ?? null;
+
 		// Vendor Name.
 		$mapped->vendor_name = $data->vendor_name ?? null;
 		if ( empty( $mapped->vendor_name ) ) {
 			$mapped->vendor_name = $data->vendorName ?? null; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 		}
+
 		// Vendor URL.
 		$mapped->vendor_url = $data->vendor_url ?? null;
 		if ( empty( $mapped->vendor_url ) ) {
 			$mapped->vendor_url = $data->vendorUrl ?? null; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 		}
+
 		// Description.
 		$mapped->description = $data->excerpt ?? null;
 		if ( empty( $mapped->description ) ) {
 			$mapped->description = $data->description ?? null;
 		}
+
 		$has_currency = ! empty( $data->currency ); // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 
 		// Is Free.
@@ -1191,17 +1277,23 @@ class WC_Admin_Addons {
 		} else {
 			$mapped->is_free = '&#36;0.00' === $data->price;
 		}
+
 		// Price.
 		if ( $has_currency ) {
 			$mapped->price = wc_price( $data->price, array( 'currency' => $data->currency ) );
 		} else {
 			$mapped->price = $data->price;
 		}
+
+		// Price suffix, e.g. "per month".
+		$mapped->price_suffix = $data->price_suffix ?? null;
+
 		// Rating.
 		$mapped->rating = $data->rating ?? null;
 		if ( null === $mapped->rating ) {
 			$mapped->rating = $data->averageRating ?? null; // phpcs:ignore WordPress.NamingConventions.ValidVariableName.UsedPropertyNotSnakeCase
 		}
+
 		// Reviews Count.
 		$mapped->reviews_count = $data->reviews_count ?? null;
 		if ( null === $mapped->reviews_count ) {
@@ -1338,7 +1430,15 @@ class WC_Admin_Addons {
 									);
 									?>
 								</span>
-								<span class="price-suffix"><?php esc_html_e( 'per year', 'woocommerce' ); ?></span>
+								<span class="price-suffix">
+									<?php
+									$price_suffix = __( 'per year', 'woocommerce' );
+									if ( ! empty( $mapped->price_suffix ) ) {
+										$price_suffix = $mapped->price_suffix;
+									}
+									echo esc_html( $price_suffix );
+									?>
+								</span>
 							<?php endif; ?>
 						</div>
 						<?php if ( ! empty( $mapped->reviews_count ) && ! empty( $mapped->rating ) ) : ?>