+++ plugins/woocommerce/includes/admin/views/html-admin-page-addons.php
@@ -6,6 +6,8 @@
  * @var string $view
  * @var object $addons
  * @var object $promotions
+ * @var array $sections
+ * @var string $current_section
  */
 
 use Automattic\WooCommerce\Admin\RemoteInboxNotifications as PromotionRuleEngine;
@@ -39,37 +41,41 @@ $current_section_name = __( 'Browse Categories', 'woocommerce' );
 	</div>
 
 	<div class="top-bar">
-		<div id="marketplace-current-section-dropdown" class="current-section-dropdown">
-			<ul>
-				<?php foreach ( $sections as $section ) : ?>
+		<ul class="marketplace-header__tabs">
+			<li class="marketplace-header__tab">
+				<a
+					class="marketplace-header__tab-link is-current"
+					href="<?php echo esc_url( admin_url( 'admin.php?page=wc-addons' ) ); ?>"
+				>
+					<?php esc_html_e( 'Browse Extensions', 'woocommerce' ); ?>
+				</a>
+			</li>
+			<li class="marketplace-header__tab">
+				<a
+					class="marketplace-header__tab-link"
+					href="<?php echo esc_url( admin_url( 'admin.php?page=wc-addons&section=helper' ) ); ?>"
+				>
 					<?php
-					if ( $current_section === $section->slug && '_featured' !== $section->slug ) {
-						$current_section_name = $section->label;
-					}
+					$count_html = WC_Helper_Updater::get_updates_count_html();
+					/* translators: %s: WooCommerce.com Subscriptions tab count HTML. */
+					echo ( sprintf( __( 'My Subscriptions %s', 'woocommerce' ), $count_html ) );
 					?>
-					<li>
-						<a
-							class="<?php echo $current_section === $section->slug ? 'current' : ''; ?>"
-							href="<?php echo esc_url( admin_url( 'admin.php?page=wc-addons&section=' . esc_attr( $section->slug ) ) ); ?>">
-							<?php echo esc_html( $section->label ); ?>
-						</a>
-					</li>
-				<?php endforeach; ?>
-			</ul>
-			<div id="marketplace-current-section-name" class="current-section-name"><?php echo esc_html( $current_section_name ); ?></div>
-		</div>
-		</div>
+				</a>
+			</li>
+		</ul>
+	</div>
 
 	<div class="wp-header-end"></div>
 
 	<div class="wrap">
 		<div class="marketplace-content-wrapper">
-			<?php if ( ! empty( $search ) && 0 === count( $addons ) ) : ?>
+			<?php require __DIR__ . '/html-admin-page-addons-category-nav.php'; ?>
+			<?php if ( ! empty( $search ) && ! is_wp_error( $addons ) && 0 === count( $addons ) ) : ?>
 				<h1 class="search-form-title">
 					<?php esc_html_e( 'Sorry, could not find anything. Try searching again using a different term.', 'woocommerce' ); ?></p>
 				</h1>
 			<?php endif; ?>
-			<?php if ( ! empty( $search ) && count( $addons ) > 0 ) : ?>
+			<?php if ( ! empty( $search ) && ! is_wp_error( $addons ) && count( $addons ) > 0 ) : ?>
 				<h1 class="search-form-title">
 					<?php // translators: search keyword. ?>
 					<?php printf( esc_html__( 'Search results for "%s"', 'woocommerce' ), esc_html( sanitize_text_field( wp_unslash( $search ) ) ) ); ?>
@@ -81,42 +87,46 @@ $current_section_name = __( 'Browse Categories', 'woocommerce' );
 					<?php WC_Admin_Addons::render_featured(); ?>
 				</div>
 			<?php endif; ?>
-			<?php if ( '_featured' !== $current_section && $addons ) : ?>
-				<?php
-				if ( ! empty( $promotions ) && WC()->is_wc_admin_active() ) {
-					foreach ( $promotions as $promotion ) {
-						WC_Admin_Addons::output_search_promotion_block( $promotion );
-					}
-				}
-				?>
-				<ul class="products">
-					<?php foreach ( $addons as $addon ) : ?>
-						<?php
-						if ( 'shipping_methods' === $current_section ) {
-							// Do not show USPS or Canada Post extensions for US and CA stores, respectively.
-							$country = WC()->countries->get_base_country();
-							if ( 'US' === $country
-								&& false !== strpos(
-									$addon->link,
-									'woocommerce.com/products/usps-shipping-method'
-								)
-							) {
-								continue;
-							}
-							if ( 'CA' === $country
-								&& false !== strpos(
-									$addon->link,
-									'woocommerce.com/products/canada-post-shipping-method'
-								)
-							) {
-								continue;
-							}
+			<?php if ( '_featured' !== $current_section ) : ?>
+				<?php if ( is_wp_error( $addons ) ) : ?>
+					<?php WC_Admin_Addons::output_empty( $addons->get_error_message() ); ?>
+				<?php else: ?>
+					<?php
+					if ( ! empty( $promotions ) && WC()->is_wc_admin_active() ) {
+						foreach ( $promotions as $promotion ) {
+							WC_Admin_Addons::output_search_promotion_block( $promotion );
 						}
+					}
+					?>
+					<ul class="products">
+						<?php foreach ( $addons as $addon ) : ?>
+							<?php
+							if ( 'shipping_methods' === $current_section ) {
+								// Do not show USPS or Canada Post extensions for US and CA stores, respectively.
+								$country = WC()->countries->get_base_country();
+								if ( 'US' === $country
+									 && false !== strpos(
+										$addon->link,
+										'woocommerce.com/products/usps-shipping-method'
+									)
+								) {
+									continue;
+								}
+								if ( 'CA' === $country
+									 && false !== strpos(
+										$addon->link,
+										'woocommerce.com/products/canada-post-shipping-method'
+									)
+								) {
+									continue;
+								}
+							}
 
-						WC_Admin_Addons::render_product_card( $addon );
-						?>
-					<?php endforeach; ?>
-				</ul>
+							WC_Admin_Addons::render_product_card( $addon );
+							?>
+						<?php endforeach; ?>
+					</ul>
+				<?php endif; ?>
 			<?php endif; ?>
 		</div>
 		<?php else : ?>