+++ plugins/woocommerce/includes/wc-conditional-functions.php
@@ -256,18 +256,6 @@ if ( ! function_exists( 'is_lost_password_page' ) ) {
 	}
 }
 
-if ( ! function_exists( 'is_ajax' ) ) {
-
-	/**
-	 * Is_ajax - Returns true when the page is loaded via ajax.
-	 *
-	 * @return bool
-	 */
-	function is_ajax() {
-		return function_exists( 'wp_doing_ajax' ) ? wp_doing_ajax() : Constants::is_defined( 'DOING_AJAX' );
-	}
-}
-
 if ( ! function_exists( 'is_store_notice_showing' ) ) {
 
 	/**
@@ -494,3 +482,31 @@ function wc_is_file_valid_csv( $file, $check_path = true ) {
 
 	return false;
 }
+
+/**
+ * Check if the current theme is a block theme.
+ *
+ * @since x.x.x
+ * @return bool
+ */
+function wc_current_theme_is_fse_theme() {
+	if ( function_exists( 'wp_is_block_theme' ) ) {
+		return (bool) wp_is_block_theme();
+	}
+	if ( function_exists( 'gutenberg_is_fse_theme' ) ) {
+		return (bool) gutenberg_is_fse_theme();
+	}
+
+	return false;
+}
+
+/**
+ * Check if the current theme has WooCommerce support or is a FSE theme.
+ *
+ * @since x.x.x
+ * @return bool
+ */
+function wc_current_theme_supports_woocommerce_or_fse() {
+	return (bool) current_theme_supports( 'woocommerce' ) || wc_current_theme_is_fse_theme();
+}
+