+++ packages/js/api-core-tests/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Added
+- Shipping Zones API Tests
+- Shipping Methods API Tests
+- Complex Order API Tests
+
 # 0.1.0
 
 - Initial/beta release
@@ -6,3 +13,5 @@
 - Coupons API Tests
 - Refunds API Tests
 - Products API Tests
+- CRUD tests for the Orders API
+- Order Search API Tests