+++ packages/js/e2e-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+## Fixed
+
+- Identified the default product category using `slug == 'uncategorized'` in `deleteAllProductCategories`
+
+## Changes
+
+- Removed `page.waitForNavigation()` from `shopper.logout()`
+
 ## Added
 
 - `utils.waitForTimeout( delay )` pause processing for `delay` milliseconds
@@ -7,6 +15,16 @@
 - Update `shopper.addToCartFromShopPage()` and `.removeFromCart()` to accept product Id or Title
 - Added `deleteAllProductAttributes()`, `deleteAllProductCategories()`, and `deleteAllProductTags()` to clean up meta data added when products are imported
 - Added `withRestApi.createProductCategory()` that creates a product category and returns the ID
+- `deleteAllProductAttributes()`, `deleteAllProductCategories()`, and `deleteAllProductTags()` to clean up meta data added when products are imported
+- `withRestApi.createProductCategory()` that creates a product category and returns the ID
+- `withRestApi.deleteCoupon()` that deletes a single coupon
+- `withRestApi.addTaxClasses()` that adds an array of tax classes if they do not exist
+- `withRestApi.addTaxRates()` that adds an array of tax rates if they do not exist
+- `clickAndWaitForSelector( buttonSelector, resultSelector, timeout )` to click a button and wait for response
+- Optional parameter `testResponse` to `withRestApi` functions that contain an `expect()`
+- `shopper.logout()` to log out the shopper account
+- `merchant.dismissOnboardingWizard()` to dismiss the onboarding wizard
+- `merchant.collapseAdminMenu()` to expand or collapse the WP admin menu
 
 # 0.1.6